запустил CRUD . победил exception при добавлении и апдейте юзера. ура! полдня на это ушло. РАБОТАЕТ. приступаем к секьюрити

+++ src/main/java/JM/Task242/dao/UserDAO.java
@@ -2,14 +2,15 @@ package JM.Task242.dao;
 
 import JM.Task242.model.User;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.List;
 
 
 public interface UserDAO {
     List<User> allUsers();
     void add(User user);
-    void delete(Long id);
+    void delete(Integer id);
     void edit(User user);
-    User getById(Long id);
+    User getById(Integer id);
     User getUserByName(String name);
 }
