+++ src/main/java/JM/Task242/service/UserService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface UserService {
     List<User> allUsers();
     void add(User user);
-    void delete(Long id);
+    void delete(Integer id);
     void edit(User user);
-    User getById(Long id);
+    User getById(Integer id);
 }
